Add files via upload
In this example, we conducted a chi-square test to determine if there is a significant association between Gender and Spending Score in the Mall Customer dataset. The results of the test showed a p-value of 0.03932, which is less than the commonly used threshold of 0.05 for statistical significance. This means that we can reject the null hypothesis that there is no association between Gender and Spending Score in the Mall Customer dataset.

Therefore, we can conclude that there is a significant association between Gender and Spending Score in the Mall Customer dataset. This indicates that the two variables are not independent and there is a relationship between a customer's gender and their spending score. We could further investigate this relationship by conducting additional analyses, such as examining the mean spending score for each gender group or visualizing the relationship using plots or charts.

Overall, this analysis tells us that there is a significant association between Gender and Spending Score in the Mall Customer data-set, and that female customers tend to have a higher spending score than male customers on average.